利用CheckStyle、findbugs等工具进行修改

+++ src/dot/DotCompiler.java
@@ -1,8 +1,11 @@
-package Dot;
-import java.io.*;
+package dot;
+
+import java.io.File;
 
 public class DotCompiler {
-	private static final String DEFAULT_CACHE_PATH = "F:\\code\\java\\SoftwareEngineeringLab\\cache";
+	//private static final String DEFAULT_CACHE_PATH = "D:\\Lab1\\bin\\Dot\\cache";
+	private static String cachePath = "/home/siyuan/cache";
+
 	public static String saveFile(String textFilePath, String dotSource) {
 		GraphViz gv = new GraphViz();
 		gv.addln(gv.start_graph());
@@ -13,8 +16,8 @@ public class DotCompiler {
 		int suffixDotIndex = textFilePath.indexOf((int)'.');
 		if (suffixDotIndex == -1)
 			imageFilePath = textFilePath;
-		else
-			imageFilePath = textFilePath.substring(0, suffixDotIndex + 1);
+		else {
+			imageFilePath = textFilePath.substring(0, suffixDotIndex + 1);}
 		File out = new File(imageFilePath + type);
 		gv.writeGraphToFile(gv.getGraph(gv.getDotSource(), type), out);
 		return imageFilePath + type;
@@ -25,17 +28,21 @@ public class DotCompiler {
 		gv.add(dotSource);
 		gv.addln(gv.end_graph());
 		String type = "png";
-		String imageFilePath = DEFAULT_CACHE_PATH + "\\" + fileName;
+		String imageFilePath = cachePath + "\\" + fileName;
 		File out = new File(imageFilePath + "." + type);
 		gv.writeGraphToFile(gv.getGraph(gv.getDotSource(), type), out);
 		return imageFilePath + "." + type;
 	}
 	public static void clearCache() {
-		File cacheDir = new File(DEFAULT_CACHE_PATH);
+		File cacheDir = new File(cachePath);
 		String[] children = cacheDir.list();
-		for (String fileName : children) {
-			File cacheFile = new File(DEFAULT_CACHE_PATH + "//" + fileName);
-			cacheFile.delete();
+		if(children != null) {
+			for (String fileName : children) {
+				File cacheFile = new File(cachePath + "//" + fileName);
+				if(cacheFile != null) {
+					cacheFile.delete();
+				}
+			}
 		}
 	}
 }
